+++ src/components/ToastPlayground/ToastPlayground.js
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import { ToastContext } from '../ToastProvider';
 import Button from '../Button';
 import ToastShelf from '../ToastShelf';
 
@@ -13,18 +14,7 @@ const VARIANT_OPTIONS = [
 ];
 
 function ToastPlayground() {
-  const [toasts, setToasts] = React.useState([
-    {
-      id: crypto.randomUUID(),
-      message: 'Oh no!',
-      variant: 'error',
-    },
-    {
-      id: crypto.randomUUID(),
-      message: 'Logged in',
-      variant: 'success',
-    },
-  ]);
+  const { createToast } = React.useContext(ToastContext);
 
   const [message, setMessage] = React.useState('');
   const [variant, setVariant] = React.useState(
@@ -34,27 +24,10 @@ function ToastPlayground() {
   function handleCreateToast(event) {
     event.preventDefault();
 
-    const nextToasts = [
-      ...toasts,
-      {
-        id: crypto.randomUUID(),
-        message,
-        variant,
-      },
-    ];
-
-    setToasts(nextToasts);
+    createToast(message, variant);
 
     setMessage('');
     setVariant(VARIANT_OPTIONS[0]);
-  }
-
-  function handleDismiss(id) {
-    const nextToasts = toasts.filter((toast) => {
-      return toast.id !== id;
-    });
-
-    setToasts(nextToasts);
   }
 
   return (
@@ -64,10 +37,7 @@ function ToastPlayground() {
         <h1>Toast Playground</h1>
       </header>
 
-      <ToastShelf
-        toasts={toasts}
-        handleDismiss={handleDismiss}
-      />
+      <ToastShelf />
 
       <form
         className={styles.controlsWrapper}
