+++ src/components/ToastShelf/ToastShelf.js
@@ -1,18 +1,17 @@
 import React from 'react';
 
+import { ToastContext } from '../ToastProvider';
 import Toast from '../Toast';
 import styles from './ToastShelf.module.css';
 
-function ToastShelf({ toasts, handleDismiss }) {
+function ToastShelf() {
+  const { toasts } = React.useContext(ToastContext);
+
   return (
     <ol className={styles.wrapper}>
       {toasts.map((toast) => (
         <li key={toast.id} className={styles.toastWrapper}>
-          <Toast
-            id={toast.id}
-            variant={toast.variant}
-            handleDismiss={handleDismiss}
-          >
+          <Toast id={toast.id} variant={toast.variant}>
             {toast.message}
           </Toast>
         </li>
